% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-module.R
\name{display}
\alias{display}
\title{Display server}
\usage{
display(input, output, session, heading, react, tooltip)
}
\arguments{
\item{input, output, session}{Usual Shiny server arguments.}

\item{heading}{Text heading.}

\item{react}{the reactive.}

\item{tooltip}{The tooltip.}
}
\description{
Display server
}
