% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parse_dates}
\alias{.parse_dates}
\title{Parse dates}
\usage{
.parse_dates(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Parse dates
}
\note{
Someone take care of this I suck with date/time stuff.
}
