% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.preproc_crawl}
\alias{.preproc_crawl}
\title{Preprocess results of webhose.io}
\usage{
.preproc_crawl(dat)
}
\arguments{
\item{dat}{Data as returned by webhoser.}
}
\description{
Preprocess results of webhose.io
}
