% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.segment}
\alias{.segment}
\title{Segment}
\usage{
.segment(data, segments, id)
}
\arguments{
\item{data}{data as returned by crawl.}

\item{segments}{segments data.frame.}

\item{id}{query id.}
}
\description{
Segment
}
