% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.segments2df}
\alias{.segments2df}
\title{Convertes segements list into a data.frame}
\usage{
.segments2df(settings)
}
\arguments{
\item{settings}{raw \code{yml} list}
}
\description{
Convertes segements list into a data.frame
}
