% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.select_segments}
\alias{.select_segments}
\title{Segment selection to query}
\usage{
.select_segments(input, type = "text")
}
\arguments{
\item{input}{Segment selected choices.}

\item{type}{Segment type.}
}
\description{
Segment selection to query
}
