% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{initial-crawl}
\alias{initial-crawl}
\alias{initial_webhose_crawl}
\title{Initial Crawl}
\usage{
initial_webhose_crawl(days = 30L, quiet = FALSE, pages = 3L)
}
\arguments{
\item{days}{Number of days from today (\code{Sys.Date()}) to craw articles.}

\item{quiet}{Whether to print helpful messages in the console (default recommended), passed to \link[webhoser]{wh_news}.}

\item{pages}{Number of pages of data to crawl, defaults to \code{3L} to crawl all 3 pages of data, set to \code{Inf} (infinite) to collect all data available under your plan (at your own rirks).}
}
\description{
Make the initial data crawl.
}
