% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load}
\alias{load}
\alias{get_articles}
\alias{get_text}
\title{Load data}
\usage{
get_articles()

get_text()
}
\description{
Load data collected with \code{\link{crawl_data}} or \code{\link{setup_auritus}}.
}
\examples{
\dontrun{data <- get_articles()}

}
