% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load}
\alias{load}
\alias{load_data}
\title{Load data}
\usage{
load_data()
}
\description{
Load data collected with \code{\link{crawl_data}} or \code{\link{setup_auritus}}.
}
\examples{
\dontrun{data <- load_data()}

}
